+++ gesipan/src/main/java/com/zzangmin/gesipan/web/service/CommentService.java
@@ -3,8 +3,8 @@ package com.zzangmin.gesipan.web.service;
 import com.zzangmin.gesipan.dao.CommentRepository;
 import com.zzangmin.gesipan.dao.PostRepository;
 import com.zzangmin.gesipan.dao.UserRepository;
-import com.zzangmin.gesipan.web.dto.CommentSaveRequest;
-import com.zzangmin.gesipan.web.dto.CommentUpdateRequest;
+import com.zzangmin.gesipan.web.dto.comment.CommentSaveRequest;
+import com.zzangmin.gesipan.web.dto.comment.CommentUpdateRequest;
 import com.zzangmin.gesipan.web.entity.Comment;
 import com.zzangmin.gesipan.web.entity.Post;
 import com.zzangmin.gesipan.web.entity.Users;
